api/requirements/base.txt
# Django and Django REST Framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment & Settings
python-dotenv==1.0.0
django-environ==0.11.2

# Authentication
django-allauth==0.57.0
dj-rest-auth==5.0.1
djangorestframework-simplejwt==5.3.0
django-axes==7.1.0

# Cache & Sessions
django-redis==5.4.0

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Utilities
Pillow==10.1.0
python-slugify==8.0.1
argon2-cffi==23.1.0

# Storage
django-storages==1.14.2
boto3==1.28.53 